Early login functionality
The general flow will need some tidying up and the file IO will all have
to be ripped out once we have a real database, but the basic login flow
is getting there. Next is setting the token as a cookie, and checking
for it before letting the user do anything.